api/main.py
"""
FastAPI メインアプリケーション
"""
from fastapi import FastAPI, HTTPException, Depends
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import List, Optional
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(
    title="株価予測API",
    description="日本の国策データを統合した株価予測システム",
    version="1.0.0"
)

# CORS設定
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)


# モデル定義
class PredictionResponse(BaseModel):
    symbol: str
    company_name: str
    prediction_date: str
    current_price: float
    target_date: str
    predicted_price: float
    confidence_score: float
    change_percentage: float
    policy_score: Optional[float] = None


class BatchPredictionRequest(BaseModel):
    symbols: List[str]
    forecast_horizon: str = "quarterly"


@app.get("/")
async def root():
    """ヘルスチェック"""
    return {"status": "ok", "message": "株価予測API稼働中"}


@app.get("/api/v1/predictions/{symbol}", response_model=PredictionResponse)
async def get_prediction(symbol: str, forecast_horizon: str = "quarterly"):
    """
    特定銘柄の予測を取得
    
    Args:
        symbol: 銘柄コード (例: 7203.T)
        forecast_horizon: 予測期間 (quarterly, monthly)
    """
    # TODO: 実際の予測ロジックを実装
    return PredictionResponse(
        symbol=symbol,
        company_name="サンプル企業",
        prediction_date="2026-02-16",
        current_price=2450.0,
        target_date="2026-05-16",
        predicted_price=2580.0,
        confidence_score=75.5,
        change_percentage=5.31,
        policy_score=68.0
    )


@app.post("/api/v1/predictions/batch")
async def batch_predictions(request: BatchPredictionRequest):
    """
    複数銘柄の予測を一括取得
    """
    # TODO: バッチ予測ロジックを実装
    predictions = []
    for symbol in request.symbols:
        predictions.append({
            "symbol": symbol,
            "predicted_price": 2500.0,
            "confidence_score": 70.0
        })
    
    return {"predictions": predictions, "request_id": "req_001"}


@app.get("/api/v1/macro-indicators")
async def get_macro_indicators():
    """
    最新のマクロ経済指標を取得
    """
    # TODO: 実際のマクロ指標データを取得
    return {
        "last_updated": "2026-02-16T14:30:00Z",
        "indicators": {
            "gdp_growth_rate": {"value": 2.1, "unit": "%", "trend": "up"},
            "cpi": {"value": 2.3, "unit": "%", "trend": "up"},
            "policy_rate": {"value": 0.25, "unit": "%", "trend": "up"},
        }
    }


@app.get("/api/v1/policy-impact")
async def get_policy_impact(days: int = 30, sector: Optional[str] = None):
    """
    最近の政策変化を取得
    """
    # TODO: 実際の政策データを取得
    return {
        "period": f"last_{days}_days",
        "policies": [
            {
                "policy_type": "gx_investment",
                "announcement_date": "2026-02-10",
                "policy_score": 82.0,
                "affected_sectors": ["renewable_energy", "automotive"],
            }
        ],
        "overall_policy_score": 75.5
    }


if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
