data_collection/boj_scraper.py
"""
日本銀行（BOJ）データスクレイパー
"""
import pandas as pd
import requests
from bs4 import BeautifulSoup
from typing import Optional
import logging

logger = logging.getLogger(__name__)


class BOJScraper:
    """日本銀行データスクレイパー"""
    
    def __init__(self):
        self.base_url = "https://www.boj.or.jp/statistics/"
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36'
        }
    
    def fetch_policy_rate(self) -> pd.DataFrame:
        """
        政策金利を取得
        
        Returns:
            政策金利データフレーム
        """
        url = f"{self.base_url}boj/other/discount/discount.htm"
        
        try:
            response = requests.get(url, headers=self.headers, timeout=30)
            response.raise_for_status()
            
            # HTMLをパース
            soup = BeautifulSoup(response.content, 'html.parser')
            
            # テーブルデータを抽出（実際のHTMLパターンに応じて調整が必要）
            tables = pd.read_html(response.content)
            
            if tables:
                df = tables[0]
                df.columns = ['date', 'rate']
                df['date'] = pd.to_datetime(df['date'], errors='coerce')
                df['rate'] = pd.to_numeric(df['rate'], errors='coerce')
                df = df.dropna()
                return df
            
            return pd.DataFrame()
            
        except Exception as e:
            logger.error(f"Failed to fetch BOJ policy rate: {e}")
            return pd.DataFrame()
    
    def fetch_jgb_10y_yield(self) -> pd.DataFrame:
        """
        10年国債利回りを取得
        
        Returns:
            10年国債利回りデータフレーム
        """
        # CSVダウンロードリンク（実際のURLに置き換える必要があります）
        url = f"{self.base_url}market/short/interest/jgbcm_all.csv"
        
        try:
            df = pd.read_csv(url, encoding='shift_jis')
            
            # カラム名を整理
            if len(df.columns) >= 2:
                df = df.iloc[:, :2]
                df.columns = ['date', 'yield']
                df['date'] = pd.to_datetime(df['date'], errors='coerce')
                df['yield'] = pd.to_numeric(df['yield'], errors='coerce')
                df = df.dropna()
                return df
            
            return pd.DataFrame()
            
        except Exception as e:
            logger.error(f"Failed to fetch JGB 10Y yield: {e}")
            return pd.DataFrame()
    
    def fetch_monetary_base(self) -> pd.DataFrame:
        """
        マネタリーベースを取得
        
        Returns:
            マネタリーベースデータフレーム
        """
        try:
            # 実際のデータソースに応じて実装
            # ここではダミーデータ
            logger.warning("Monetary base fetching not fully implemented")
            return pd.DataFrame()
            
        except Exception as e:
            logger.error(f"Failed to fetch monetary base: {e}")
            return pd.DataFrame()
