requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0

# Data collection
yfinance==0.2.32
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.26.3

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
torch==2.1.2
lightgbm==4.1.0

# API
fastapi==0.109.0
uvicorn==0.27.0
pydantic-settings==2.1.0

# Caching
redis==5.0.1

# Monitoring and logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Code quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Utilities
requests-cache==1.1.1
tenacity==8.2.3
